Async-Operations with CancellationToken #283
Performs an async read from a stream with a CancellationToken.
This CancellationToken must be passed to every method as the user can use
 different Timeouts.

The new method plc.GetDefaultCancellationToken() provides a new token,
which is cancelled after half a second.

To maintain API-Compatibility it might be worth considering an overload where
the default token is used.