For ReadStruct/WriteStruct, add ordering of fields by declaration order.
This should fix #135 , but it does require users to add [StructLayout(LayoutKind.Sequential)] attribute to their classes.

This implementation throws if that is not the case, so will break existing code. I am not sure how lenient the library intends to be here or not.

Also fix a parsing error for DWord in ReadStruct, which mixed up endianes.