DateTimeLong and Class to support DateTime and Strings
The DateTimeLong appears to have been updated by another member during the same period I had checked out.

For Class support, I am including the features which we are actively using for our processes which include DateTime conversion based on method injection of PlcType. This should have no impact on released versions as the PLC Type is defined when opening the connection. The implementation of strings within a class is very common in .NET, but there is no simple way to inject a length. For that reason I have established the default string size (equal to maximum string size 254 char).